Display a chart for selected stock
Implemented new feature to display a chart for the selected stock. Users can now view the performance of the stock over various time intervals such as 1 day, 5 days, 1 month, 6 months, 1 year, 5 years, or the maximum available data.
Additionally, the stock logo will be shown next to the stock symbol in the quote section.